Refactored existing code to make scripting functionality easier to implement
